+++ src/Xpring/xpring-client.ts
@@ -3,7 +3,7 @@ import { BigInteger } from 'big-integer'
 import XrpPayIdClientInterface, {
   XRPPayIDClientInterface,
 } from '../PayID/xrp-pay-id-client-interface'
-import XRPClientInterface from '../XRP/xrp-client-interface'
+import { XRPClientInterface } from '../XRP/xrp-client-interface'
 import XpringError, { XpringErrorType } from './xpring-error'
 import XrplNetwork, { XRPLNetwork } from '../Common/xrpl-network'
 import { XRPPayIDClient } from '../PayID/xrp-pay-id-client'
@@ -30,6 +30,7 @@ export default class XpringClient {
    * @param xrpClient An XRP Client used to interact with the XRP Ledger protocol.
    * @throws A XpringError if the networks of the inputs do not match.
    */
+  // TODO(keefertaylor): Support new xrp client.
   constructor(
     payIdClient: XRPPayIDClientInterface | XrpPayIdClientInterface,
     private readonly xrpClient: XRPClientInterface,