+++ test/XRP/xrp-transaction-type-proto-conversion.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai'
 
-import XRPAccountSet from '../../src/XRP/model/xrp-account-set'
+import { XRPAccountSet } from '../../src/XRP/model/xrp-account-set'
 
 import {
   testAccountSetProtoAllFields,